February Matches (#228)
* Tumbleweed DLLs

* Match: ECSH_Cup (+ ECSH_Shrine docs)

* ECSH_Shrine docs small edit

* ECSH_Shrine update func signature

* Match: CRspellstone

* Sound enum update

* Match: DLL 280 SidekickBall

* SidekickBall syms edit

* Match: DLL 59 minimap

* Match: DLL 322 coldWaterControl

* Match: DLL 323 waterfallControl

* Match: DLL 324 curve

* Match: DLL 344 SpellOverlay

* Match: DLL 357 ExplodeAnimator

* Match: DLL 418 DFriverflow

* Match: DLL 494 NWtreebridge

* Match: DLL 511 SHboulder

* Match: DLL 603 WMvein

* Match: DLL 606 WMwallpower

* Match: DLL 617 WMvconsole

* Match: DLL 654 textblock

* DLL 658 PointBack progress

* Match: DLL 659 DBtrigger

* Match: DLL 660 DBspike

* Match: DLL 683 DBsupshield

* Match: DLL 685 DBbonedust

* Match: DLL 684 DBdustgeezer

* Match: DLL 691 DBfakehorizon

* Match: DLL 605 WLworm

* Match: DLL 742 DRexplodeDoor

* Match: DLL 431 DFwhirlpool

* Match: DLL 562 DIM2icicle

* Match: dll_297_func_1544 (thanks to Bloodangel!)

* Match: DLL 297 Scarab

* Gamebit update for nonmatching

* Match: DLL 638 DFPLevelControl

* Match: DLL 644 DFPfloorbar

* Match: DLL 644 DFPfloorbar

* Inventory item header update

* Object header notes

* ObjDef header notes

* Small cleanup/docs: segment_326A0

* Match: DLL 523 SCcollectables (and related ObjDef changes)

* Struct update: DLL 324 curve

* Added SideFoodTypes enum

* Merging textable enums

+++ src/segment_53F00.c
@@ -720,7 +720,7 @@ void func_80054DF8(UnkFunc80051D68Arg3* arg0, UnkFunc80051D68Arg3* arg1, u8 arg2
     }
 }
 
-u8 func_8005509C(Object *arg0, f32* arg1, f32* arg2, s32 arg3, Unk80027934* arg4, u8 arg5) {
+s32 func_8005509C(Object *arg0, f32* arg1, f32* arg2, s32 arg3, Unk80027934* arg4, u8 arg5) {
     Unk8005341C* var_s3;
     Unk8005341C* target;
     f32 spA0[4 * 3];
